Check if NFC is enabled in onResume activity
This essentially forces the user to enable NFC to even be able to enter
the passport number etc., but you can't really do anything with the app
anyways. So I figured it was better to just forcefully request the user
to enable it?

It shows the help text in a toast message.

Fixes #49